HDPI-6424: Add Wiremock mapping for Fee Register and remove FakeFeeService
