+++ src/main/resources/application-dev.yaml
@@ -1,6 +1,3 @@
-fees:
-  mock: true
-
 spring:
   flyway:
     locations: classpath:db/migration,classpath:db/testdata
