+++ wiremock/mappings/fee-service/FEE0380-enforcement-warrant.json
@@ -1,0 +1,38 @@
+{
+  "request": {
+    "urlPath": "/fees-register/fees/lookup",
+    "method": "GET",
+    "queryParameters": {
+      "service": {
+        "equalTo": "civil money claims"
+      },
+      "jurisdiction1": {
+        "equalTo": "civil"
+      },
+      "jurisdiction2": {
+        "equalTo": "county court"
+      },
+      "channel": {
+        "equalTo": "default"
+      },
+      "event": {
+        "equalTo": "enforcement"
+      },
+      "keyword": {
+        "equalTo": "WarrantPossorDelivery"
+      }
+    }
+  },
+  "response": {
+    "status": 200,
+    "headers": {
+      "Content-Type": "application/json"
+    },
+    "jsonBody": {
+      "code": "FEE0380",
+      "description": "Issue of a warrant of possession/warrant of delivery",
+      "version": 5,
+      "fee_amount": 148.00
+    }
+  }
+}
